+++ mysensors/gateway_serial.py
@@ -23,6 +23,7 @@ class SerialGateway(Gateway, threading.Thread):
         threading.Thread.__init__(self)
         Gateway.__init__(self, event_callback, persistence,
                          persistence_file, protocol_version)
+        self.lock = threading.Lock()
         self.serial = None
         self.port = port
         self.baud = baud
@@ -50,7 +51,7 @@ class SerialGateway(Gateway, threading.Thread):
         except serial.SerialException:
             _LOGGER.error('Unable to connect to %s', self.port)
             return False
-        self._schedule_save_sensors()
+        self.persistence.schedule_save_sensors()
         return True
 
     def disconnect(self):
@@ -66,8 +67,8 @@ class SerialGateway(Gateway, threading.Thread):
         """Stop the background thread."""
         _LOGGER.info('Stopping thread')
         self._stop_event.set()
-        if self.scheduled_save is not None:
-            self.scheduled_save.cancel()
+        if self.persistence.scheduled_save is not None:
+            self.persistence.scheduled_save.cancel()
 
     def run(self):
         """Background thread that reads messages from the gateway."""
@@ -103,7 +104,7 @@ class SerialGateway(Gateway, threading.Thread):
             self.fill_queue(self.logic, (string,))
         self.disconnect()  # Disconnect after stop event is set
         if self.persistence:
-            self._save_sensors()
+            self.persistence.save_sensors()
 
     def send(self, message):
         """Write a Message to the gateway."""
