+++ mysensors/mysensors.py
@@ -1,7 +1,9 @@
 """Convenience module for backwards compatibility."""
 # flake8: noqa: F401
 # pylint: disable=unused-import
-from mysensors import ChildSensor, Gateway, Message, Sensor
+from mysensors import Gateway
 from mysensors.gateway_mqtt import MQTTGateway
 from mysensors.gateway_serial import SerialGateway
 from mysensors.gateway_tcp import TCPGateway
+from mysensors.message import Message
+from mysensors.sensor import ChildSensor, Sensor
